Add/poc client api using a hook for creating the poll (#58)
The idea is we use a custom hook which we then use in the edit block for handling bidirectional communication with the API.

The hook does 3 things:

    When it is mounted it tries to load the poll from the API in case an id already exists.
    It exposes a function setOutboundChanges that allows us to indicate that changes happened from the user on the block.
    It exposes a function maybeSyncQueuedChanges. It is called on every block paint and is responsible for syncing stuff to and from the api.

A bit more detail on what maybeSyncQueuedChanges does:

    if we have a poll in state ensure the id is saved on the block.
    Do nothing if syncing is already taking place
    if we have outbound changes (from the block by the user), send them out and reset outbound counter
    if we have inbound changes (from the api), call "sync complete" with the current version of the attributes. Reset inbound counter.

Testing:

make client

On your local setup, create a post. Open the inspector and look at console.

Add a poll block. A post request is issued and a poll is created. Poll id is set.
Edit poll, more requests are fired but not all the time, even if you keep going super-fast about it!
if there are any pending changes, you will see a final post when any pending update request is done