Ensure any inbound changes are applied before sending any outbound ones. (#91)
This will ensure that anything that any entities with ids will be
updated in the block as well.

Testing:

create a poll block, and immediately write an answer, hit return and
write another one. if you observe the network requests, the first one
will have an id and it will be saved. Then the new request will include
the first answer id in it, so no new answer will be created for the
first one.