Store the API key as well as the user code.
The API key is used to talk to the API server too so we need to store
that as well.
The setup function has been modified so it checks for that, and then
uses the key if found to get a user code. If that fails then we ask the
user to get another API key as the cached one may have been deleted on
the Crowdsignal.com site.