Plugin-side changes for using client-defined ids for polls and answers (#110)
Plugin-side changes for using client-defined ids for polls and answers

Introduces a new gateway class for handling client -> platform poll data
associations.

Refactors copy-pasted code in the gateway.

On the frontend side of things, it removes setAttributesAndFlagChange calls, as well as the need for the client=based sync logic (which can be completely removed in a separate pr). Additionally, the way we get the ids to vote on is altered.